+++ src/PimEnterprise/Bundle/WorkflowBundle/Doctrine/ORM/Repository/PublishedProductRepository.php
@@ -40,6 +40,22 @@ class PublishedProductRepository extends ProductRepository implements PublishedP
     /**
      * {@inheritdoc}
      */
+    public function findOneByOriginalProductId($originalProductId)
+    {
+        return $this->findOneBy(['originalProduct' => $originalProductId]);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function findOneByVersionId($versionId)
+    {
+        return $this->findOneBy(['version' => $versionId]);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
     public function findByOriginalProducts(array $originalProducts)
     {
         $originalIds = [];